+++ lib/types.ts
@@ -1,4 +1,5 @@
 /* eslint-disable max-len */
+import type { RequestInit } from 'undici';
 
 /**
  * The options used by Open Graph Scraper